+++ Sources/APIManager/APIService.swift
@@ -35,7 +35,13 @@ public protocol APIService {
     /// ```
     static var baseURL: String { get }
 
-    /// The `HTTPHeaders` to be sent alongside the `APIRequest`s made by the endpoints in this `APIService`. No headers are sent by default, so any and all headers must be specified here.
+    /// The `HTTPHeaders` to be sent alongside the `APIRequest`s made by the endpoints in this `APIService`.
+    static var parameters: HTTPParameters? { get }
+
+    /// The `HTTPHeaders` to be sent alongside the `APIRequest`s made by the endpoints in this `APIService`.
+    static var body: HTTPBody? { get }
+
+    /// The `HTTPHeaders` to be sent alongside the `APIRequest`s made by the endpoints in this `APIService`.
     static var headers: HTTPHeaders? { get }
 
     /// Offers an optional point for an `APIService` to determine what HTTPReponse status codes are valid. The default implementation marks any response with a status code in the range 200..<300 as valid.
@@ -46,6 +52,15 @@ public protocol APIService {
 
 /// Default implementations of APIService components
 public extension APIService {
+    /// The `HTTPHeaders` to be sent alongside the `APIRequest`s made by the endpoints in this `APIService`.
+    static var parameters: HTTPParameters? { return nil }
+
+    /// The `HTTPHeaders` to be sent alongside the `APIRequest`s made by the endpoints in this `APIService`.
+    static var body: HTTPBody? { return nil }
+    
+    /// The `HTTPHeaders` to be sent alongside the `APIRequest`s made by the endpoints in this `APIService`.
+    static var headers: HTTPHeaders? { return nil }
+
     /// Default implementation of validate that marks any response with a status code in the range 200..<300 as valid.
     public static func validate(statusCode: Int) throws {
         if !(200..<300).contains(statusCode) {