+++ Sources/APIManager/HTTPBody.swift
@@ -1,9 +1,0 @@
-//
-//  HTTPBody.swift
-//  APIManager
-//
-//  Created by Rauhul Varma on 4/21/17.
-//  Copyright © 2017 Rauhul Varma. All rights reserved.
-//
-
-public typealias HTTPBody = [String: Any]